Added threshold flag
Added a flag for thresholding. This is the maximum percentage
difference between assemblages for all the types that will be tolerated
in adding to a solution. If not specified, then all assemblages will be
tested. If you specify -threshold=0.05 only those assemblages that have
a maximum type frequency of less than 0.05 will be added to the ends.